+++ packages/labs/analyzer/src/lib/javascript/classes.ts
@@ -72,6 +72,28 @@ export const getClassDeclaration = (
   });
 };
 
+const getIsReadonlyForNode = (
+  node: ts.Node,
+  analyzer: AnalyzerInterface
+): boolean => {
+  const {typescript} = analyzer;
+  if (typescript.isPropertyDeclaration(node)) {
+    return (
+      node.modifiers?.some((mod) =>
+        typescript.isReadonlyKeywordOrPlusOrMinusToken(mod)
+      ) ||
+      typescript
+        .getJSDocTags(node)
+        .some((tag) => tag.tagName.text === 'readonly')
+    );
+  } else if (typescript.isStatement(node)) {
+    return typescript
+      .getJSDocTags(node)
+      .some((tag) => tag.tagName.text === 'readonly');
+  }
+  return false;
+};
+
 /**
  * Returns the `fields` and `methods` of a class.
  */
@@ -84,10 +106,42 @@ export const getClassMembers = (
   const staticFieldMap = new Map<string, ClassField>();
   const methodMap = new Map<string, ClassMethod>();
   const staticMethodMap = new Map<string, ClassMethod>();
+  const accessors = new Map<string, {get?: ts.Node; set?: ts.Node}>();
   declaration.members.forEach((node) => {
     // Ignore non-implementation signatures of overloaded methods by checking
     // for `node.body`.
-    if (typescript.isMethodDeclaration(node) && node.body) {
+    if (typescript.isConstructorDeclaration(node) && node.body) {
+      // TODO(bennypowers): We probably want to see if this matches what TypeScript considers a field initialization.
+      // Maybe instead of iterating through the constructor statements, we walk the body looking for any
+      // assignment expression so that we get ones inside of if statements, in parenthesized expressions, etc.
+      //
+      // Also, this doesn't cover destructuring assignment.
+      //
+      // This is ok for now because these are rare ways to "declare" a field,
+      // especially in web components where you shouldn't have constructor parameters.
+      node.body.statements.forEach((node) => {
+        if (
+          typescript.isExpressionStatement(node) &&
+          typescript.isBinaryExpression(node.expression) &&
+          node.expression.operatorToken.kind ===
+            typescript.SyntaxKind.EqualsToken &&
+          typescript.isPropertyAccessExpression(node.expression.left) &&
+          node.expression.left.expression.kind ===
+            typescript.SyntaxKind.ThisKeyword
+        ) {
+          const name = node.expression.left.name.getText();
+          fieldMap.set(
+            name,
+            new ClassField({
+              name,
+              type: getTypeForNode(node.expression.right, analyzer),
+              privacy: getPrivacy(typescript, node),
+              readonly: getIsReadonlyForNode(node, analyzer),
+            })
+          );
+        }
+      });
+    } else if (typescript.isMethodDeclaration(node) && node.body) {
       const info = getMemberInfo(typescript, node);
       const name = node.name.getText();
       (info.static ? staticMethodMap : methodMap).set(
@@ -99,7 +153,10 @@ export const getClassMembers = (
         })
       );
     } else if (typescript.isPropertyDeclaration(node)) {
-      if (!typescript.isIdentifier(node.name)) {
+      if (
+        !typescript.isIdentifier(node.name) &&
+        !typescript.isPrivateIdentifier(node.name)
+      ) {
         analyzer.addDiagnostic(
           createDiagnostic({
             typescript,
@@ -115,6 +172,7 @@ export const getClassMembers = (
       }
 
       const info = getMemberInfo(typescript, node);
+
       (info.static ? staticFieldMap : fieldMap).set(
         node.name.getText(),
         new ClassField({
@@ -122,10 +180,33 @@ export const getClassMembers = (
           default: node.initializer?.getText(),
           type: getTypeForNode(node, analyzer),
           ...parseNodeJSDocInfo(node, analyzer),
+          readonly: getIsReadonlyForNode(node, analyzer),
+        })
+      );
+    } else if (typescript.isAccessor(node)) {
+      const name = node.name.getText();
+      const _accessors = accessors.get(name) ?? {};
+      if (typescript.isGetAccessor(node)) _accessors.get = node;
+      else if (typescript.isSetAccessor(node)) _accessors.set = node;
+      accessors.set(name, _accessors);
+    }
+  });
+  for (const [name, {get, set}] of accessors) {
+    if (get ?? set) {
+      fieldMap.set(
+        name,
+        new ClassField({
+          name,
+          type: getTypeForNode((get ?? set)!, analyzer),
+          privacy: getPrivacy(typescript, (get ?? set)!),
+          readonly: !!get && !set,
+          // TODO(bennypowers): derive from getter?
+          // default: ???
+          // TODO(bennypowers): reflect, etc?
         })
       );
     }
-  });
+  }
   return {
     fieldMap,
     staticFieldMap,