+++ packages/labs/analyzer/test-files/js/properties/element-a.js
@@ -39,8 +39,25 @@ export class ElementA extends LitElement {
 
   [unsupportedPropertyName] = '';
 
+  /** @type {number} */
+  get getterOnly() {
+    return 0;
+  }
+
+  /** @type {number} */
+  get accessorPair() {
+    return 0;
+  }
+  set accessorPair(_) {
+    void 0;
+  }
+
+  /** @readonly */
+  readonlyField = 0;
+
   constructor() {
     super();
+    this.constructorAssignOnly = 0;
     this.notDecorated = '';
     this.noOptionsString = '';
     this.noOptionsNumber = 42;