+++ packages/labs/react/src/create-component.ts
@@ -48,15 +48,27 @@ export type ReactWebComponent<
 // lifecycle methods or allow user to explicitly provide props.
 type ElementProps<I> = Partial<Omit<I, keyof HTMLElement>>;
 
-// Acceptable props to the React component. Omit keyof E from HTMLAttributes to
-// prefer provided event handler mapping over React's built-in event handler
-// props.
+// Child type compatible with both React and Preact. It's based on
+// `React.ReactNode` but using `JSX.Element` allows`VNode` to be acceptable in
+// Preact projects.
+type Child =
+  | JSX.Element
+  | React.ReactPortal
+  | string
+  | number
+  | boolean
+  | undefined;
+
+// Acceptable props to the React component.
 type ComponentProps<I, E extends EventNames = {}> = Omit<
   React.HTMLAttributes<I>,
-  keyof E
+  // Omit keyof E to prefer provided event handler mapping over React's
+  // built-in event handler props.
+  | keyof E
+  // Omit children to replace with our own that's compatible with Preact.
+  | 'children'
 > &
-  ElementProps<I> &
-  EventListeners<E>;
+  EventListeners<E> & {children?: Child | Child[]} & ElementProps<I>;
 
 /**
  * Type used to cast an event name with an event type when providing the